+++ src/glassesTools/recording.py
@@ -15,7 +15,7 @@ class Recording:
     source_directory            : pathlib.Path  = ""
     working_directory           : pathlib.Path  = ""
     start_time                  : Timestamp     = Timestamp(0)
-    duration                    : int           = None
+    duration                    : float         = None
     eye_tracker                 : EyeTracker    = EyeTracker.Unknown
     eye_tracker_name            : str           = ""    # name to show if eye_tracker is EyeTracker.Generic
     project                     : str           = ""
